requirements.txt
accelerate==1.2.0
adbc_driver_manager==1.3.0
adbc_driver_postgresql==1.3.0
adbc_driver_sqlite==1.3.0
ads==0.12.6
aerospike_vector_search==3.0.1
ain==0.0.2
aiodns==3.2.0
airbyte_cdk==6.9.2
aleph_alpha_client==8.0.0
alibabacloud_ha3engine_vector==1.1.10
alibabacloud_tea_util==0.3.13
anywidget==0.9.13
apex==0.9.10dev
apify_client==1.8.1
AppKit==0.2.8
arango==0.2.1
arcgis==2.4.0
argilla==2.5.0
arize==7.29.3
arthurai==3.32.3
astrapy==1.5.2
async_timeout==5.0.1
atheris==2.3.0
atlassian==0.0.0
audiocraft==1.3.0
av==14.0.1
azure_storage==0.37.0
azureml==0.2.7
baidubce==0.8.3
banana_dev==6.3.0
bibtexparser==1.4.2
bigdl==2.4.0
bilibili_api==9.1.0
bokeh==3.6.2
boto3==1.35.78
botocore==1.35.78
brotli==1.1.0
brotlicffi==1.1.0.0
cassio==0.1.10
causal_conv1d==1.5.0.post8
chardet==5.2.0
checks==0.2
chromadb==0.5.23
clarifai_grpc==10.11.0
classy_vision==0.7.0
clearml==1.16.5
clickhouse_connect==0.8.9
cloudpathlib==0.20.0
cloudpickle==3.1.0
codecarbon==2.8.1
colorama==0.4.6
comet_llm==2.2.7
comet_ml==3.47.4
compressed_tensors==0.8.0
ConfigParser==7.1.0
confluent_kafka==2.6.1
cryptography==44.0.0
ctypes_snappy==1.03
cudf==0.6.1.post1
curio==1.6
cx_Oracle==8.3.0
cycler==0.12.1
Cython==3.0.11
dagshub==0.3.46
dall_e==0.1
dask_expr==1.1.20
datadog_api_client==2.30.0
deepeval==2.0.5
defusedxml==0.7.1
dgml_utils==0.3.4
dingodb==0.0.18
distributed==2024.12.0
doctran==0.0.14
docutils==0.21.2
docx2txt==0.8
dria==0.0.99
dvclive==3.48.0
dynamodb_encryption_sdk==3.3.0
e2b==1.0.5
einops==0.8.0
elevenlabs==1.13.4
email_validator==2.2.0
esprima==4.0.1
essentia==2.1b6.dev1177
eval_type_backport==0.2.0
exceptiongroup==1.2.2
extract_msg==0.52.0
fairseq==0.12.2
fairseq2==0.2.1
fake_useragent==2.0.3
fastai==2.7.18
fastapi==0.115.6
faster_whisper==1.1.0
fastparquet==2024.11.0
fbgemm_gpu==1.0.0
feedparser==6.0.11
firebase_admin==6.6.0
fitz==0.0.1.dev2
flash_attn==2.7.2.post1
flashrank==0.2.9
flatdict==4.0.1
flytekit==1.14.0
Foundation==0.1.0a0.dev1
fqdn==1.5.1
ftfy==6.3.1
fugashi==1.4.0
g2p_en==2.1.0
galore_torch==1.0
gdown==5.2.0
geopandas==1.0.1
gguf==0.10.0
gitdb_speedups==0.1.0
gliner==0.2.13
gluonnlp==0.10.0
google_api_python_client==2.154.0
google_auth_oauthlib==1.2.1
googlemaps==4.10.0
gptcache==0.1.44
gpudb==7.2.2.3
gql==3.5.0
gradientai==1.13.1
gradio==5.8.0
gradio_client==1.5.1
grandalf==0.8
greenlet==3.1.1
grokadamw==0.1.2
grpc==1.0.0
guardrails==2.0.0
gunicorn==23.0.0
h2==4.1.0
h5py==3.12.1
haiku==0.01
hdbcli==2.23.24
hf_transfer==0.1.8
hologres_vector==0.0.10
HTMLParser==0.0.2
hypothesis==6.122.3
ibis==3.3.0
ibm_watsonx_ai==1.1.25
importlib_metadata==8.0.0
importlib_resources==6.4.5
infinity_client==0.0.73
infinity_emb==0.0.73
InquirerPy==0.3.4
InstructorEmbedding==1.0.1
intel_extension_for_pytorch==2.5.0
intel_extension_for_transformers==1.4.2
ipadic==1.0.0
ipython==8.12.3
ipywidgets==8.1.5
isal==1.7.1
isoduration==20.11.0
jaguardb_http_client==3.4.2.3
javelin_sdk==0.2.19
jax==0.4.37
jaxlib==0.4.36
jieba==0.42.1
jnius==1.1.0
joblib==1.4.2
JPype1==1.5.1
jq==1.8.0
kdbai_client==1.5.0
kenlm==0.2.0
keras_nlp==0.18.1
kerchunk==0.2.7
keybert==0.8.5
keyring==25.5.0
label_studio_sdk==1.0.8
laion_clap==1.1.6
langchain_anthropic==0.3.0
langchain_astradb==0.5.2
langchain_aws==0.2.9
langchain_chroma==0.1.4
langchain_cohere==0.3.3
langchain_elasticsearch==0.3.0
langchain_experimental==0.3.3
langchain_fireworks==0.2.5
langchain_google_genai==2.0.6
langchain_google_vertexai==2.0.8
langchain_groq==0.2.1
langchain_huggingface==0.1.2
langchain_mistralai==0.2.3
langchain_mongodb==0.2.0
langchain_neo4j==0.1.1
langchain_ollama==0.2.1
langchain_openai==0.2.12
langchain_pinecone==0.2.0
langchain_postgres==0.0.12
langchain_qdrant==0.2.0
langchain_together==0.2.0
langchainhub==0.1.21
langsmith_pyo3==0.1.0rc4
lark==1.2.2
laser_encoders==0.0.2
librosa==0.10.2.post1
liger_kernel==0.5.1
listparser==0.20
llmlingua==0.2.2
llmonitor==0.0.51
lomo_optim==0.1.1
lorax==0.94.86
lxml==5.3.0
lz4==4.3.3
lzmaffi==0.3.0
mamba_ssm==2.2.4
mambapy==1.2.0
manticoresearch==4.0.0
mask==1.0.0
Mask2Former==0.1.5
matplotlib==3.9.3
mediawikiapi==1.2
megatron==0.5.1
metal_sdk==2.5.1
mistral_common==1.5.1
ml_dtypes==0.5.0
mlx_lm==0.20.2
modelscope==1.21.0
modin==0.32.0
motor==3.6.0
msal==1.31.1
mtrand==0.1
MultiScaleDeformableAttention==1.0
mwparserfromhell==0.6.6
mwxml==0.3.4
mxnet==1.9.1
natten==0.17.3
neptune==1.13.0
nest_asyncio==1.6.0
newspaper==0.1.0.7
nlu==5.4.1
nomic==3.3.3
nucliadb_models==6.1.0.post2518
nucliadb_protos==6.1.0.post2518
numexpr==2.10.2
O365==2.0.38
obs==0.0.0
oci==2.140.0
odfpy==1.4.1
odps==3.5.1
olefile==0.47
onnxruntime==1.20.1
openapi_pydantic==0.5.0
openpyxl==3.1.5
optimum==1.23.3
optuna==4.1.0
outcome==1.3.0.post0
paramiko==3.5.0
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
phonemizer==3.3.0
pickle5==0.0.12
pinecone_text==0.9.0
pkgutil_resolve_name==1.3.10
plotly==5.24.1
polars==1.17.1
praw==7.8.1
pretty_midi==0.2.10
promptlayer==1.0.31
psutil==6.1.0
psychicapi==0.8.4
py==1.11.0
py3nvml==0.2.7
pyairtable==3.0.1
pycares==4.5.0
pycocotools==2.0.8
pycryptodome==3.21.0
pyctcdecode==0.5.0
pycurl==7.45.3
pydub==0.25.1
pyepsilla==0.3.10
pygit2==1.16.0
pygraphviz==1.14
PyInstaller==6.11.1
PyJWT==2.10.1
pymemcache==4.0.0
pymilvus==2.5.0
pymongo==4.10.1
pyodide==0.0.2
pyOpenSSL==24.3.0
pyowm==3.3.0
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
pyperf==2.8.1
pyppeteer==2.0.0
PyQt4==4.11.4
PyQt5==5.15.11
PyQt6==6.7.1
PySide6==6.8.1
pyspark==3.5.3
pysqlcipher3==1.2.0
pysrt==1.1.2
pytesseract==0.3.13
pytest==8.3.4
pythainlp==5.0.4
python-decouple==3.8
python_calamine==0.3.1
python_Levenshtein==0.26.1
pyTigerGraph==1.8.3
pytorch_lightning==2.4.0
pytorch_quantization==2.2.1
pytube==15.0.0
pyxlsb==1.0.10
qcloud_cos==3.3.6
qdrant_client==1.12.1
qianfan==0.4.12.1
qtpy==2.4.2
rank_bm25==0.2.2
rank_llm==0.20.2
rapidfuzz==3.10.1
rapidocr_onnxruntime==1.4.2
ray==2.40.0
rdflib==7.1.1
reportlab==4.2.5
requests_kerberos==0.15.0
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rfc3987==1.3.8
rhoknp==1.7.0
rizaio==0.3.0
rjieba==0.1.11
rockset==2.1.2
rspace_client==2.5.0
runhouse==0.0.36
s3fs==2024.10.0
s3prl==0.4.16
sacremoses==0.1.1
schedulefree==1.4
scipy==1.14.1
scrapingant_client==2.1.0
selenium==4.27.1
semantic_version==2.10.0
sentencepiece==0.2.0
sets==0.3.2
setuptools==75.6.0
setuptools_scm==8.1.0
sigopt==8.8.3
simplejson==3.19.3
simsimd==6.2.1
slack_sdk==3.33.5
smbprotocol==1.15.0
socksio==1.0.0
sodapy==2.2.0
soundfile==0.12.1
speechkit==2.2.2
Sphinx==8.1.3
sqlcipher3==0.5.4
sqlite_vec==0.1.6
sqlite_vss==0.1.2
sseclient==0.0.27
stackapi==0.3.1
starlette==0.41.3
steamship==2.17.34
steamspypi==1.1.1
sudachipy==0.6.9
svglib==1.5.1
sympy==1.13.1
t5x==0.0.0
tables==3.10.1
takeoff_client==0.6.0
telethon==1.38.1
tensorboardX==2.6.2.2
tensorflow_probability==0.25.0
tensorflow_text==2.18.0
tensorstore==0.1.69
textractor==0.1.2
tf2onnx==1.16.1
tf_keras==2.18.0
thirdai==0.9.24
threadpoolctl==3.5.0
tigrisdb==1.0.0b6
timescale_vector==0.0.7
timm==1.0.12
tokenizer==3.4.5
torch_npu==2.4.0
torch_xla==2.5.1
torchao==0.7.0
torchaudio==2.5.1
torchvision==0.20.1
traitlets==5.14.3
trajectory==0.1.1
tree_sitter==0.23.2
tree_sitter_languages==1.10.2
trubrics==1.6.2
tweepy==4.14.0
ujson==5.10.0
unidic==1.1.0
unidic_lite==1.0.8
upstash_ratelimit==1.1.0
upstash_vector==0.6.0
uri_template==1.3.0
urllib3_secure_extra==0.1.0
uroman==1.3.1.1
uvicorn==0.32.1
uvloop==0.21.0
vearch_cluster==0.0.2
vegafusion==2.0.1
vissl==0.1.6
watchdog==6.0.0
webcolors==24.11.1
wget==3.2
whylogs==1.6.4
wikibase_rest_api_client==0.2.1
wolframalpha==5.1.3
writerai==3.4.0
xarray==2024.11.0
xformers==0.0.28.post3
xinference_client==1.0.1
xlrd==2.0.1
xlsxwriter==3.2.0
xmlrpclib==1.0.1
xmltodict==0.14.2
yfinance==0.2.50
youtube_search==2.1.2
youtube_transcript_api==0.6.3
yt_dlp==2024.12.6
zep_python==2.0.2
zstandard==0.23.0
